+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Initial Setup: Initial setup the project and possible dependencies & relevants setup on September 23, 2024 -> **0.0.0+1** by Luke
 - Initial Development: Place the app fundamentals on September 23, 2024 -> **0.0.1** by Luke
+- Initial Development: Setup Firebase for the app, added logger system, added the permission system + notification system & setting system (language), and other improvements on September 27, 2024 -> **0.0.2** by Luke